requirements.txt
fastapi[all]
numpy
opencv-python
imutils
Pillow
scikit-image